+++ semcore/core/src/utils/use/useFocusLock.ts
@@ -183,7 +183,7 @@ const useFocusLockHook = (
     if (returnFocusTo === 'auto' && autoTriggerRef.current) {
       const autoTrigger = autoTriggerRef.current;
       setTimeout(() => {
-        if (focusMastersStackCount !== focusMastersStack.length) {
+        if (focusMastersStackCount !== focusMastersStack.length || autoFocus === 'enforced') {
           setFocus(autoTrigger, trapNode);
         }
       }, 0);
@@ -218,6 +218,11 @@ const useFocusLockHook = (
 
     if (autoFocus) {
       setTimeout(() => {
+        if (autoFocus === 'enforced' && focusableChildren.length === 0 && focusMaster) {
+          trapRef.current?.focus();
+          return;
+        }
+
         setFocus(
           trapRef.current!,
           typeof returnFocusTo === 'object' ? returnFocusTo?.current : document.body,