+++ semcore/core/CHANGELOG.md
@@ -2,6 +2,19 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [16.0.3] - 2025-07-02
+
+### Added
+
+- `useScrollBarWidth` hook.
+- Container focus if there aren't focusable children while `autoFocus='enforced'`
+
+## [16.0.2] - 2025-06-23
+
+### Added
+
+- `focusVisible` option to `setFocus` util method with value from last interaction type.
+
 ## [16.0.1] - 2025-06-16
 
 ### Added