+++ semcore/inline-input/CHANGELOG.md
@@ -2,11 +2,17 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
-## [16.1.4] - 2025-07-01
+## [16.1.5] - 2025-07-07
 
 ### Fixed
 
 - Inline input and its actions are focusable even when disabled.
+
+## [16.1.4] - 2025-07-04
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/flex-box` [16.0.3 ~> 16.0.4], `@semcore/icon` [16.2.2 ~> 16.2.3], `@semcore/tooltip` [16.0.3 ~> 16.0.4], `@semcore/button` [16.0.3 ~> 16.0.4], `@semcore/base-components` [16.1.0 ~> 16.1.1]).
 
 ## [16.1.3] - 2025-06-23